Refactor applicaton route/controller
In the current-user component, the load function now peeks for the
user record, since the record should already exist.
The methods for loading settings, the current user and the user
preferences were moved to the application controller.
The methods are partially async now. That way, it is no longer necessary
to use promises as return values.
With the loadUserAndSettings task, we can make sure that the navbar
will only be displayed when both the user was authenticated and all
settings and preferences were loaded. Before, the navbar was diaplyed
next to the login form for a moment before moving to the visualization
route.
This commit also fixes the bug that was causing the user settings not to
be displayed when entering a direct url, as the component was initialized
before the settings in the application route were loaded.
